python/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --allow-unsafe ./python/requirements.txt
#

#manual add
pandas-stubs==1.2.0.58
openpyxl==1.1.0
pyxlsb==1.0.9

##

bcrypt==3.2.0
    # via paramiko
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.6
    # via requests
cryptography==3.4.8
    # via paramiko
idna==3.2
    # via requests
influxdb==5.3.1
    # via -r ./python/requirements.txt
msgpack==1.0.2
    # via influxdb
paramiko==2.10.1
    # via -r ./python/requirements.txt
prettytable==2.2.0
    # via -r ./python/requirements.txt
pycparser==2.20
    # via cffi
pynacl==1.4.0
    # via paramiko
python-dateutil==2.8.2
    # via influxdb
pytz==2021.1
    # via influxdb
requests==2.26.0
    # via
    #   -r ./python/requirements.txt
    #   influxdb
six==1.16.0
    # via
    #   bcrypt
    #   influxdb
    #   pynacl
    #   python-dateutil
urllib3==1.26.7
    # via requests
wcwidth==0.2.5
    # via prettytable
